requirements.txt
streamlit
langchain
langchain-community
langchain-groq
sentence-transformers
chromadb
scikit-learn
python-dotenv
pypdf